Factor out request method <span>, format code
