+++ frontend/src/pages/RequestsPage/RequestsPage.tsx
@@ -15,9 +15,9 @@ import { DataTable } from "@/components/ui/DataTable";
 import { Button } from "@/components/ui/button";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { type MizuTrace, useMizuTraces } from "@/queries";
+import { cn } from "@/utils";
 import { Row, getPaginationRowModel } from "@tanstack/react-table";
 import { columns } from "./columns";
-import { cn } from "@/utils";
 
 type LevelFilter = "all" | "error" | "warning" | "info" | "debug";
 
@@ -80,11 +80,10 @@ export function RequestsPage() {
   }, [queryClient]);
 
   return (
-    <Tabs defaultValue="all" className={cn(
-      "py-4 px-2",
-      "sm:px-4 sm:py-4",
-      "md:px-6 md:py-6",
-    )}>
+    <Tabs
+      defaultValue="all"
+      className={cn("py-4 px-2", "sm:px-4 sm:py-4", "md:px-6 md:py-6")}
+    >
       <div className="flex items-center">
         <TabsList>
           <TabsTrigger value="all">All</TabsTrigger>
